issue #316 Handling UPDATEs in the framework of DIVBIP

+++ extensions/bip/src/edu/ucsc/dbtune/bip/core/InumQueryPlanDesc.java
@@ -107,7 +107,7 @@ public class InumQueryPlanDesc implements QueryPlanDesc, Serializable
     }
     
     @Override
-    public double getStatementWeight()
+    public int getStatementWeight()
     {
         return stmt.getStatementWeight();
     }
